Implemented pagination for endpoint and error handling with invaid pagesize/pageindex
- Changed VM.PageIndex and VM.PageSize to int for ease in coding
- Move DB variable in KOLLogic to local to avoid nil pointer dereference
- Implemented pagination in reference to https://gorm.io/docs/scopes.html
- Error handling in case of invalid pagesize/pageindex, return status BadRequest (400)

+++ BE/ViewModels/KOLViewModel.go
@@ -5,8 +5,8 @@ import "wan-api-kol-event/DTO"
 type KolViewModel struct {
 	Result       string        `json:"result"`       //Result : success, unsuccess
 	ErrorMessage string        `json:"errorMessage"` // The query error
-	PageIndex    int64         `json:"pageIndex"`    //The page index
-	PageSize     int64         `json:"pageSize"`     //The page size
+	PageIndex    int           `json:"pageIndex"`    //The page index
+	PageSize     int           `json:"pageSize"`     //The page size
 	Guid         string        `json:"guid"`
 	TotalCount   int64         `json:"totalCount"`
 	KOL          []*DTO.KolDTO `json:"kol"`
